+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/callbacks/PaymentElementCallbacks.kt
@@ -8,7 +8,6 @@ import com.stripe.android.paymentelement.CreateIntentWithConfirmationTokenCallba
 import com.stripe.android.paymentelement.EmbeddedPaymentElement
 import com.stripe.android.paymentelement.EmbeddedPaymentElement.RowSelectionBehavior.Companion.getInternalRowSelectionCallback
 import com.stripe.android.paymentelement.ExperimentalAnalyticEventCallbackApi
-import com.stripe.android.paymentelement.ExperimentalCustomPaymentMethodsApi
 import com.stripe.android.paymentelement.PreparePaymentMethodHandler
 import com.stripe.android.paymentelement.ShopPayPreview
 import com.stripe.android.paymentelement.TapToAddPreview
@@ -19,7 +18,6 @@ import com.stripe.android.paymentsheet.ExternalPaymentMethodConfirmHandler
 import com.stripe.android.paymentsheet.ShopPayHandlers
 
 @OptIn(
-    ExperimentalCustomPaymentMethodsApi::class,
     ExperimentalAnalyticEventCallbackApi::class,
     ShopPayPreview::class,
     SharedPaymentTokenSessionPreview::class,